Remove temperature timer, cleanup code
Instead of using a timer it simpler to just delay the task for a second.
Since temperature reading is used for controlling the oven it was also
moved to control.

Unnecessary logs during hardware initialization were also removed.
